// apps/web/components/ui/button.d.ts
import React from 'react';
export declare const Button: ({ children, onClick, variant, size, className, disabled, type }: {
    children: React.ReactNode;
    onClick?: () => void;
    variant?: "default" | "secondary" | "outline" | "ghost" | "destructive";
    size?: "default" | "sm" | "lg";
    className?: string;
    disabled?: boolean;
    type?: "button" | "submit" | "reset";
}) => React.JSX.Element;
//# sourceMappingURL=button.d.ts.map